% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floratus.R
\docType{data}
\name{floratus}
\alias{floratus}
\title{Dataframe with 80,000 occurrence records for Tuscany region}
\format{
A data frame with 80,000 rows and 5 variables:
\describe{
  \item{Taxon}{Taxonomic Identity}
  \item{Long}{Longitude, in WGS84}
  \item{Lat}{Latitude, in WGS84}
  \item{uncertainty}{radius of the uncertainty buffer from the occurrence geographic position, in metres}
  \item{year}{the year of the occurrence}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
floratus
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
