% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrfi_example.R
\name{load_mrfi_example}
\alias{load_mrfi_example}
\title{Load MRFI Example Raster}
\usage{
load_mrfi_example()
}
\value{
A terra SpatRaster object with MRFI values.
}
\description{
Loads the example Map of Relative Floristic Ignorance (MRFI) raster.
}
\details{
This MRFI was generated using the ignorance_map() function from the ignobioR package,
based on floristic occurrence data with spatial and temporal uncertainty.
Higher values indicate areas with greater floristic knowledge gaps.
}
\examples{
\dontrun{
mrfi <- load_mrfi_example()
terra::plot(mrfi)
}
}
