% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/park.R
\docType{data}
\name{park}
\alias{park}
\title{Boundaries of the Migliarino-San Rossore-Massaciuccoli Regional Park (Italy)}
\format{
A polygon of class SpatialPolygonsDataFrame having CRS "+init=epsg:4326"
\describe{
 }
}
\usage{
park
}
\description{
Qui ci va la descrizione!
}
\keyword{datasets}
