% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsuitablezone.R
\docType{data}
\name{unsuitablezone}
\alias{unsuitablezone}
\title{Water areas of Migliarino-San Rossore-Massaciuccoli Regional Park (Italy)}
\format{
A polygon of class SpatialPolygonsDataFrame having CRS "+init=epsg:4326"
\describe{
}
}
\usage{
unsuitablezone
}
\description{
Areas in which records cannot occur due to a given reason (according to the user decision).
}
\keyword{datasets}
