% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floratus.R
\docType{data}
\name{floratus}
\alias{floratus}
\title{Floristic occurrence records for Tuscany region}
\format{
A data frame with 80,000 rows and 5 variables:
\describe{
  \item{Taxon}{Taxonomic Identity}
  \item{Long}{Longitude (in WGS84)}
  \item{Lat}{Latitude (in WGS84)}
  \item{uncertainty}{radius of the uncertainty buffer rising from the occurrence geographic position, in metres}
  \item{year}{the year of the occurrence}
}
}
\source{
\url{http://bot.biologia.unipi.it/wpb/italia/index.html; https://www.gbif.org/}
}
\usage{
floratus
}
\description{
A dataset containing 80,000 florisitic occurrence records for Tuscany
}
\keyword{datasets}
