% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignorance_map.R
\name{ignorance_map}
\alias{ignorance_map}
\title{Mapping Floristic Ignorance}
\usage{
ignorance_map(data_flor, site, year_study, excl_areas, cellsize, CRS.new, tau)
}
\arguments{
\item{site, }{excl_areas, tau, cellsize}
}
\value{
A map
}
\description{
This function map occurence data taking into account spatial and temporal uncertainty
}
\examples{
\dontrun{ignorance_map()}
}
