% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignorance_map.R
\name{ignorance_map}
\alias{ignorance_map}
\title{Ignorance map}
\usage{
ignorance_map(
  data_flor,
  site,
  year_study = NULL,
  excl_areas = NULL,
  CRS.new = 3035,
  tau,
  cellsize
)
}
\arguments{
\item{data_flor}{dataframe having 5 columns, namely ‘Taxon’ (species identity), ‘Long’ (longitude coordinates), ‘Lat’ (latitude coordinates), ‘uncertainty’ (radius of uncertainty, in metres), and ‘year’ (year of the record)}

\item{site}{a layer object of class ‘SpatialPolygonsDataFrame’ representing the study area, having CRS: +init=epsg:4326}

\item{year_study}{the present-year in which you perform the analysis}

\item{excl_areas}{a layer object of class ‘SpatialPolygonsDataFrame’ to delimit certainly unsuitable areas adjacent or within the study area, having CRS: +init=epsg:4326}

\item{CRS.new}{he new Coordinate Reference System. Note: must be in XXXXXXXXXXXXlll}

\item{tau}{percentual value of taxa loss in 100 years time-span (see below for further details)}

\item{cellsize}{the resolution of the ignorance map (in meters)}
}
\value{
A list with 4 objects:
\itemize{
 \item{"MFI"}{ The Map of Floristic Ignorance}
 \item{"RICH"}{ The corresponding map computed without taking into account spatial and temporal uncertainties}
 \item{"uncertainties"}{ The corresponding map computed without taking into account spatial and temporal uncertainties}
 \item{"Statistics"}{ A table summarising the settings used to draft the Map of Florstic Ignorance}
}
}
\description{
A list of species potentially occurring within a study site, in which a probability of occurrence is computed for every taxon
}
\examples{
\dontrun{
data(floratus)
data(park)
data(unsuitablezone)

ignorance_map(data_flor = floratus, excl_areas = unsuitablezone, site = park, tau = 20, cellsize = 2000)
}
}
