% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/park.R
\docType{data}
\name{park}
\alias{park}
\title{Boundaries of the Migliarino-San Rossore-Massaciuccoli Regional Park (Tuscany, Italy)}
\format{
A polygon of class 'SpatialPolygonsDataFrame' having CRS "+init=epsg:4326"
\describe{The study area of the example
 }
}
\usage{
park
}
\description{
The study area of the example provided
}
\keyword{datasets}
