% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_list.R
\name{virtual_list}
\alias{virtual_list}
\title{Draft the virtual floristic list of a given territory}
\usage{
virtual_list(
  data_flor,
  site,
  year_study,
  excl_areas = NULL,
  CRS.new,
  tau,
  upperlimit
)
}
\arguments{
\item{site, }{tau}
}
\value{
A dataframe
}
\description{
Bla bla bla
}
\examples{
\dontrun{virtual_list()}
}
