% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_list.R
\name{virtual_list}
\alias{virtual_list}
\title{Draft the virtual floristic list (VFL) of the study area}
\usage{
virtual_list(
  data_flor,
  site,
  year_study = NULL,
  excl_areas = NULL,
  CRS.new = 3035,
  tau,
  upperlimit = 20
)
}
\arguments{
\item{data_flor}{dataframe having 5 columns, namely ‘Taxon’ (species identity), ‘Long’ (longitude coordinates), ‘Lat’ (latitude coordinates), ‘uncertainty’ (radius of uncertainty, in metres), and ‘year’ (year of the record)}

\item{site}{a layer object of class ‘SpatialPolygonsDataFrame’ representing the study area, having CRS: +init=epsg:4326}

\item{year_study}{the present-year in which you perform the analysis}

\item{excl_areas}{a layer object of class ‘SpatialPolygonsDataFrame’ to delimit certainly unsuitable areas adjacent or within the study area, having CRS: +init=epsg:4326}

\item{CRS.new}{he new Coordinate Reference System. Note: must be in XXXXXXXXXXXX}

\item{tau}{YYYYYYYYYYYYY}

\item{upperlimit}{VVVVVVVVVVVVV}
}
\value{
A .csv file (comma-separated values) file having N columns, namely ‘Taxon’, YYYYYYYYYYYYYY
}
\description{
This function map occurrence data taking into account spatial and temporal uncertainty of each record
}
\examples{
\dontrun{
data(datashort)
data(site)
data(exclareas)

virtual_list(data_flor=datashort, site=site, excl_areas=exclareas, CRS.new=3035, tau=30, upperlimit=25)}
}
