% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_list.R
\name{virtual_list}
\alias{virtual_list}
\title{Draft the virtual floristic list of a given territory}
\usage{
virtual_list(
  data_flor,
  site,
  year_study,
  excl_areas = NULL,
  CRS.new,
  tau,
  upperlimit
)
}
\arguments{
\item{data_flor, }{site, year_study, excl_areas, CRS.new, tau, upperlimit
tau}
}
\value{
A .csv file (comma-separated values) file having N columns, namely ‘Taxon’, YYYYYYYYYYYYYY
}
\description{
A list of species potentially occurring within a study site, in which a probability of occurrence is computed for every taxon
}
\examples{
\dontrun{virtual_list()}
}
